split up model files
