+++ corehq/warehouse/models/__init__.py
@@ -1,0 +1,12 @@
+from corehq.warehouse.models.dimensions import (
+    UserDim,
+    GroupDim,
+    LocationDim,
+    DomainDim,
+    UserLocationDim,
+    UserGroupDim,
+)
+
+from corehq.warehouse.models.meta import (
+    FactTableState,
+)
